Agent/requirements.txt
python-dotenv
requests
yfinance
yahooquery
openai
openai-agents
streamlit
matplotlib
finnhub-python
aiohttp
